backend/requirements.txt
Flask
Flask-Cors
textblob
requests
beautifulsoup4
google-api-python-client
google-generativeai
PyPDF2
python-dotenv